Recurrence_id on activities + admin panel changes
Added a recurrence_id on the activity model that can be used to signify one-time exceptions for one of the parent_activity's fields.
This field is already validated (e.g. a parent activity cannot have a parent activity itself).

The "view on site" button in the admin panel is now enabled for activities that support it.

Added a page that will list all one-time exceptions of an activity (view was based on the "History" view in the admin panel). This button appears next to the "History" button.