Added CompoundActivity
The CompoundActivity is used to handle a group of activites as a single one. It's not only for organisational purposes, but also for handling one-time exceptions (E.g. to modify the event's start time for just a single occurence)

The relevant migrations (including a data migration) have been added for this.